Serializer: added support for std::pair. BinaryArchive a few changes.

+++ src/cpputils/serialization/binaryarchive.h
@@ -2,7 +2,7 @@
 #include <stdexcept>
 #include <vector>
 
-constexpr uint64_t INITIAL_CAPACITY = 128ull;
+constexpr size_t INITIAL_CAPACITY = 128ull;
 
 class BinaryArchive {
  public:
@@ -27,33 +27,33 @@ class BinaryArchive {
 
   // Constructs instance and reserves memory for internal storage. By
   // default, storage initial capacity equals 'INITIAL_CAPACITY' constant value.
-  BinaryArchive(uint64_t capacity = INITIAL_CAPACITY);
+  BinaryArchive(size_t capacity = INITIAL_CAPACITY);
 
   // Copies 'size' amount of data from the internal storage to the memory
   // location, pointed by 'data' argument. Affects reading position, increasing
   // it by 'size' argument value. Throws exception if 'size' argument value
   // exceeds internal storage bound, considering last reading position.
-  void Read(unsigned char* const data, uint64_t size);
+  void Read(unsigned char* const data, uint32_t size);
 
   // Copies 'size' amount of data from the memory location, pointed by 'data'
   // argument to internal storage. Affects writing position, increasing it by
   // 'size' argument value.
   // Expands internal storage size while inserting new data.
   // Throws exception if internal storage expanding failed.
-  void Write(const unsigned char* const data, uint64_t size);
+  void Write(const unsigned char* const data, uint32_t size);
 
   // Returns current reading position depending on seek direction.
-  uint64_t GetReadPosition(SeekDirection seekDir = SeekDirection::BEGIN) const;
+  int32_t GetReadPosition(SeekDirection seekDir = SeekDirection::BEGIN) const;
 
   // Sets new value of reading position depending on seek direction.
-  void SetReadPosition(uint64_t pos,
+  void SetReadPosition(int32_t pos,
                        SeekDirection seekDir = SeekDirection::BEGIN);
 
   // Returns current writing position depending on seek direction.
-  uint64_t GetWritePosition(SeekDirection seekDir = SeekDirection::BEGIN) const;
+  int32_t GetWritePosition(SeekDirection seekDir = SeekDirection::BEGIN) const;
 
   // Sets new value of writing position depending on seek direction.
-  void SetWritePosition(uint64_t pos,
+  void SetWritePosition(int32_t pos,
                         SeekDirection seekDir = SeekDirection::BEGIN);
 
   // Sets size of internal storage to 0.
@@ -62,7 +62,7 @@ class BinaryArchive {
   void Clear();
 
   // Returns size of internal storage.
-  uint64_t GetSize() const;
+  size_t GetSize() const;
 
   // Returns data pointer to first element of internal storage.
   const unsigned char* GetDataPointer() const;
@@ -70,18 +70,18 @@ class BinaryArchive {
  private:
   enum class SeekType { READ, WRITE };
 
-  uint64_t calculateNewPosition(uint64_t pos, SeekType seekType,
-                                SeekDirection seekDir) const;
-  uint64_t calculateCurrentPosition(SeekType seekType,
-                                    SeekDirection seekDir) const;
+  int32_t calculateNewPosition(int32_t pos, SeekType seekType,
+                               SeekDirection seekDir) const;
+  int32_t calculateCurrentPosition(SeekType seekType,
+                                   SeekDirection seekDir) const;
 
  private:
   // Internal storage.
   std::vector<unsigned char> data;
 
   // Writing position.
-  uint64_t ppos;
+  int32_t ppos;
 
   // Reading position.
-  uint64_t gpos;
+  int32_t gpos;
 };
